essentiels/requirements-ocr.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
easyocr==1.7.0
opencv-python-headless==4.8.1.78
Pillow==9.5.0
numpy==1.24.3
